k8s/apps/order-api/requirements.txt
flask
psycopg2-binary
gunicorn
redis
prometheus_client
